@@ -33,7 +33,6 @@ import fr.umlv.smalljs.ast.Expr.LocalVarAssignment;
 import fr.umlv.smalljs.ast.Expr.MethodCall;
 import fr.umlv.smalljs.ast.Expr.New;
 import fr.umlv.smalljs.ast.Expr.Return;
-import fr.umlv.smalljs.ast.VoidVisitor;
 import fr.umlv.smalljs.rt.Failure;
 import fr.umlv.smalljs.rt.JSObject;
 
@@ -71,10 +70,6 @@ public class InstrRewriter {
 		}
 	}
 
-	private InstrRewriter(Dictionary dict, InstrBuffer buffer, JSObject globalEnv) {
-		this.visitor = createVisitor(buffer, dict, globalEnv);
-	}
-
 	public static JSObject createFunction(Optional<String> name, List<String> parameters, Block body, Dictionary dict, JSObject globalEnv) {
 		var env = JSObject.newEnv(null);
 
@@ -85,8 +80,7 @@ public class InstrRewriter {
 		visitVariable(body, env);
 
 		var buffer = new InstrBuffer();
-		var rewriter = new InstrRewriter(dict, buffer, globalEnv);
-		rewriter.rewrite(body, env);
+		visit(body, env, buffer, dict, globalEnv);
 		buffer.emit(CONST).emit(encodeDictObject(UNDEFINED, dict));
 		buffer.emit(RET);
 
@@ -105,185 +99,202 @@ public class InstrRewriter {
 	}
 
 	private static void visitVariable(Expr expr, JSObject env) {
-		VARIABLE_VISITOR.visit(expr, env);
-	}
-
-	private static final VoidVisitor<JSObject> VARIABLE_VISITOR = new VoidVisitor<JSObject>()
-			.when(Block.class, (block, env) -> {
-				for (var instr : block.instrs()) {
+		switch (expr) {
+			case Block block -> {
+				for (Expr instr : block.instrs()) {
 					visitVariable(instr, env);
 				}
-			}).when(Literal.class, (literal, env) -> {
+			}
+			case Literal literal -> {
 				// do nothing
-			}).when(FunCall.class, (funCall, env) -> {
+			}
+			case FunCall funCall -> {
 				// do nothing
-			}).when(LocalVarAssignment.class, (localVarAssignment, env) -> {
+			}
+			case LocalVarAssignment localVarAssignment -> {
 				if (localVarAssignment.declaration()) {
 					env.register(localVarAssignment.name(), env.length());
 				}
-			}).when(LocalVarAccess.class, (localVarAccess, env) -> {
+			}
+			case LocalVarAccess localVarAccess -> {
 				// do nothing
-			}).when(Fun.class, (fun, env) -> {
+			}
+			case Fun fun -> {
 				// do nothing
-			}).when(Return.class, (_return, env) -> {
+			}
+			case Return _return -> {
 				// do nothing
-			}).when(If.class, (_if, env) -> {
+			}
+			case If _if -> {
 				visitVariable(_if.trueBlock(), env);
 				visitVariable(_if.falseBlock(), env);
-			}).when(New.class, (_new, env) -> {
-			  // do nothing
-			}).when(FieldAccess.class, (fieldAccess, env) -> {
-			  // do nothing
-			}).when(FieldAssignment.class, (fieldAssignment, env) -> {
-			  // do nothing
-			}).when(MethodCall.class, (methodCall, env) -> {
-			  // do nothing
-			});
-
-	private void rewrite(Expr expr, JSObject env) {
-		visitor.visit(expr, env);
+			}
+			case New _new -> {
+				// do nothing
+			}
+			case FieldAccess fieldAccess -> {
+				// do nothing
+			}
+			case FieldAssignment fieldAssignment -> {
+				// do nothing
+			}
+			case MethodCall methodCall -> {
+				// do nothing
+			}
+			default -> throw new AssertionError("unknown case " + expr.getClass());
+		};
 	}
 
-	public static VoidVisitor<JSObject> createVisitor(InstrBuffer buffer, Dictionary dict, JSObject globalEnv) {
-		var visitor = new VoidVisitor<JSObject>();
-		visitor.when(Block.class, (block, env) -> {
-			// for each expression of the block
-			for (var expr : block.instrs()) {
-				// visit the expression
-				visitor.visit(expr, env);
-				// if the expression is an instruction (i.e. return void)
-				if (!(expr instanceof Instr)) {
-					// ask to top the top of the stack
-					buffer.emit(POP);
+	private static void visit(Expr expr, JSObject env, InstrBuffer buffer, Dictionary dict, JSObject globalEnv) {
+		switch (expr) {
+			case Block block -> {
+				// for each expression of the block
+				for (var instr : block.instrs()) {
+					// visit the expression
+					visit(instr, env, buffer, dict, globalEnv);
+					// if the expression is an instruction (i.e. return void)
+					if (!(instr instanceof Instr)) {
+						// ask to top the top of the stack
+						buffer.emit(POP);
+					}
 				}
 			}
-		}).when(Literal.class, (literal, env) -> {
-			throw new UnsupportedOperationException("TODO Literal");
-			// get the literal value
-			//var value = ...
-			// test if it's a positive integers
-			//if (value instanceof Integer && ((Integer) value) >= 0) {
+			case Literal literal -> {
+				throw new UnsupportedOperationException("TODO Literal");
+				// get the literal value
+				//var literalValue = ...
+				// test if it's a positive integers
+				//if (literalValue instanceof Integer value && value >= 0) {
 				// emit a small int
 				//buffer.emit(...).emit(...);
-			//} else {
+				//} else {
 				// emit a dictionary object
 				//buffer.emit(...).emit(...);
-			//}
-		}).when(FunCall.class, (funCall, env) -> {
-			throw new UnsupportedOperationException("TODO FunCall");
-			// visit the qualifier
-			//visitor.visit(...);
-			// emit undefined
-			//buffer.emit(...).emit(...)
-			// visit all arguments
-			//for (var arg : funCall.args()) {
-			//	visitor.visit(arg, env);
-			//}
-			// emit the funcall
-			//buffer.emit(...).emit(...);
-		}).when(LocalVarAccess.class, (localVarAccess, env) -> {
-			throw new UnsupportedOperationException("TODO LocalVarAccess");
-			// get the local variable name
-			//var name = ...
-			// find if there is a local variable in the environment with the name
-			//var slotOrUndefined = env.lookup(...);
-			//if (slotOrUndefined == UNDEFINED) {
+				//}
+			}
+			case FunCall funCall -> {
+				throw new UnsupportedOperationException("TODO FunCall");
+				// visit the qualifier
+				//visitor.visit(...);
+				// emit undefined
+				//buffer.emit(...).emit(...)
+				// visit all arguments
+				//for (var arg : funCall.args()) {
+				//	visitor.visit(...);
+				//}
+				// emit the funcall
+				//buffer.emit(...).emit(...);
+			}
+			case LocalVarAccess localVarAccess -> {
+				throw new UnsupportedOperationException("TODO LocalVarAccess");
+				// get the local variable name
+				//var name = ...
+				// find if there is a local variable in the environment with the name
+				//var slotOrUndefined = env.lookup(...);
+				//if (slotOrUndefined == UNDEFINED) {
 				// emit a lookup with the name
 				//buffer.emit(...).emit(...);
-			//} else {
+				//} else {
 				// load the local variable with the slot
 				//buffer.emit(...).emit(...);
-			//}
-		}).when(LocalVarAssignment.class, (localVarAssignment, env) -> {
-			throw new UnsupportedOperationException("TODO LocalVarAssignment");
-			// visit the expression
-			// visitor.visit(...);
-			// get the local variable name
-			//var name = ...
-			// find if there is a local variable in the env from the name
-			//var slotOrUndefined = env.lookup(...);
-			//if (slotOrUndefined == UNDEFINED) {
-			//	throw new Failure("unknown local variable " + name);
-			//}
-			// emit a store at the variable slot
-			//buffer.emit(...).emit(...);
-		}).when(Fun.class, (fun, env) -> {
-			throw new UnsupportedOperationException("TODO Fun");
-			// create a JSObject function
-			///var function = createFunction(fun.name(), fun.parameters(), fun.body(), dict, globalEnv);
-			// emit a const on the function
-			//buffer.emit(...).emit(...);
-			// if the name is present emit a code to register the function in the global environment
-			//fun.name().ifPresent(name -> {
+				//}
+			}
+			case LocalVarAssignment localVarAssignment -> {
+				throw new UnsupportedOperationException("TODO LocalVarAssignment");
+				// visit the expression
+				// visitor.visit(...);
+				// get the local variable name
+				//var name = ...
+				// find if there is a local variable in the env from the name
+				//var slotOrUndefined = env.lookup(...);
+				//if (slotOrUndefined == UNDEFINED) {
+				//	throw new Failure("unknown local variable " + name);
+				//}
+				// emit a store at the variable slot
+				//buffer.emit(...).emit(...);
+			}
+			case Fun fun -> {
+				throw new UnsupportedOperationException("TODO Fun");
+				// create a JSObject function
+				///var function = createFunction(fun.name(), fun.parameters(), fun.body(), dict, globalEnv);
+				// emit a const on the function
+				//buffer.emit(...).emit(...);
+				// if the name is present emit a code to register the function in the global environment
+				//fun.name().ifPresent(name -> {
 				//buffer.emit(DUP);
 				//buffer.emit(...).emit(...);
-			//});
-		}).when(Return.class, (_return, env) -> {
-			throw new UnsupportedOperationException("TODO Return");
-			// emit a visit of the expression
-			//visitor.visit(...);
-			// emit a RET
-			//buffer.emit(RET);
-		}).when(If.class, (_if, env) -> {
-			throw new UnsupportedOperationException("TODO If");
-			// visit the condition
-			//visitor.visit(...);
-			// emit a JUMP_IF_FALSE and a placeholder
-			//var falsePlaceHolder = buffer.emit(JUMP_IF_FALSE).placeholder();
-			// visit the true block
-			//visitor.visit(...);
-			// emit a goto with another placeholder
-			//var endPlaceHolder = buffer.emit(GOTO).placeholder();
-			// patch the first placeholder
-			//buffer.patch(..., buffer.label());
-			// visit the false block
-			//visitor.visit(...);
-			// patch the second place holder
-			//buffer.patch(..., buffer.label());
-		}).when(New.class, (_new, env) -> {
-			throw new UnsupportedOperationException("TODO New");
-		  // create a JSObject class
-			//var clazz = JSObject.newObject(null);
-			// loop over all the field initializations
-			//_new.initMap().forEach((fieldName, expr) -> {
-				// register the field name with the right slot
-			  //clazz.register(...);
-			  // visit the initialization expression
+				//});
+			}
+			case Return _return -> {
+				throw new UnsupportedOperationException("TODO Return");
+				// emit a visit of the expression
 				//visitor.visit(...);
-			//});
-			// emit a NEW with the class
-			//buffer.emit(...).emit(...);
-		}).when(FieldAccess.class, (fieldAccess, env) -> {
-			throw new UnsupportedOperationException("TODO FieldAccess");
-			// visit the receiver
-			//visitor.visit(...);
-			// emit a GET with the field name
-			//buffer.emit(...).emit(...);
-		}).when(FieldAssignment.class, (fieldAssignment, env) -> {
-			throw new UnsupportedOperationException("TODO FieldAssignment");
-			// visit the receiver
-			//visitor.visit(...);
-		  // visit the expression
-		  //visitor.visit(...);
-			// emit a PUT with the field name
-			//buffer.emit(...).emit(...);
-		}).when(MethodCall.class, (methodCall, env) -> {
-			throw new UnsupportedOperationException("TODO MethodCall");
-		  // visit the receiver
-		  //visitor.visit(...);
-		  // emit a DUP, get the field name and emit a SWAP of the qualifier and the receiver
-		  //buffer.emit(DUP);
-		  //buffer.emit(...).emit(...);
-		  //buffer.emit(SWAP);
-		  // visit all arguments
-		  //for (var arg : methodCall.args()) {
-		    //visitor.visit(...);
-			//}
-		  // emit the funcall
-			//buffer.emit(...).emit(...);
-		});
-		return visitor;
+				// emit a RET
+				//buffer.emit(RET);
+			}
+			case If _if -> {
+				throw new UnsupportedOperationException("TODO If");
+				// visit the condition
+				//visitor.visit(...);
+				// emit a JUMP_IF_FALSE and a placeholder
+				//var falsePlaceHolder = buffer.emit(JUMP_IF_FALSE).placeholder();
+				// visit the true block
+				//visitor.visit(...);
+				// emit a goto with another placeholder
+				//var endPlaceHolder = buffer.emit(GOTO).placeholder();
+				// patch the first placeholder
+				//buffer.patch(..., buffer.label());
+				// visit the false block
+				//visitor.visit(...);
+				// patch the second place holder
+				//buffer.patch(..., buffer.label());
+			}
+			case New _new -> {
+				throw new UnsupportedOperationException("TODO New");
+				// create a JSObject class
+				//var clazz = JSObject.newObject(null);
+				// loop over all the field initializations
+				//_new.initMap().forEach((fieldName, expr) -> {
+				// register the field name with the right slot
+				//clazz.register(...);
+				// visit the initialization expression
+				//visitor.visit(...);
+				//});
+				// emit a NEW with the class
+				//buffer.emit(...).emit(...);
+			}
+			case FieldAccess fieldAccess -> {
+				throw new UnsupportedOperationException("TODO FieldAccess");
+				// visit the receiver
+				//visitor.visit(...);
+				// emit a GET with the field name
+				//buffer.emit(...).emit(...);
+			}
+			case FieldAssignment fieldAssignment -> {
+				throw new UnsupportedOperationException("TODO FieldAssignment");
+				// visit the receiver
+				//visitor.visit(...);
+				// visit the expression
+				//visitor.visit(...);
+				// emit a PUT with the field name
+				//buffer.emit(...).emit(...);
+			}
+			case MethodCall methodCall -> {
+				throw new UnsupportedOperationException("TODO MethodCall");
+				// visit the receiver
+				//visitor.visit(...);
+				// emit a DUP, get the field name and emit a SWAP of the qualifier and the receiver
+				//buffer.emit(DUP);
+				//buffer.emit(...).emit(...);
+				//buffer.emit(SWAP);
+				// visit all arguments
+				//for (var arg : methodCall.args()) {
+				//visitor.visit(...);
+				//}
+				// emit the funcall
+				//buffer.emit(...).emit(...);
+			}
+			default -> throw new AssertionError("unknown expr " + expr.getClass());
+		}
 	}
-
-	private final VoidVisitor<JSObject> visitor;
 }
