@@ -3,9 +3,9 @@ package fr.umlv.smalljs.stackinterp;
 import fr.umlv.smalljs.rt.JSObject;
 
 public interface TagValues {
-	// every values are stored as 32 bits integer, boolean, small ints, constant (dictionary object) and reference
+	// every value are stored as 32 bits integer, boolean, small ints, constant (dictionary object) and reference
 	// the suffix indicates the kind of value
-	//    1 -> small ints or boolean (TRUE/FALSE)
+	//    1 -> small integers (SmallInt) or boolean (TRUE/FALSE)
 	//   10 -> dictionary index
 	//   00 -> reference 
 	
@@ -39,26 +39,25 @@ public interface TagValues {
   }
 
   static Object decodeAnyValue(int tagValue, Dictionary dict, int[] heap) {
-    if (TagValues.isSmallInt(tagValue)) {
-      return TagValues.decodeSmallInt(tagValue);
+    if (isSmallInt(tagValue)) {
+      return decodeSmallInt(tagValue);
     }
-    if (TagValues.isReference(tagValue)) {
-      var ref = TagValues.decodeReference(tagValue);
+    if (isReference(tagValue)) {
+      var ref = decodeReference(tagValue);
       var clazz = (JSObject) decodeDictObject(heap[ref], dict);
       return clazz.mirror(offset -> decodeAnyValue(heap[ref + OBJECT_HEADER_SIZE + (int)offset], dict, heap));
     }
-    return TagValues.decodeDictObject(tagValue, dict);
+    return decodeDictObject(tagValue, dict);
   }
   static int encodeAnyValue(Object object, Dictionary dict) {
-  	int smallInt;
-    if (object instanceof Integer && (smallInt = (Integer)object) >= 0) {
-      return TagValues.encodeSmallInt(smallInt);
+  	if (object instanceof Integer smallInt && smallInt >= 0) {
+      return encodeSmallInt(smallInt);
     }
-    return TagValues.encodeDictObject(object, dict);
+    return encodeDictObject(object, dict);
   }
 
-  int TRUE = TagValues.encodeSmallInt(1);
-  int FALSE = TagValues.encodeSmallInt(0);
+  int TRUE = encodeSmallInt(1);
+  int FALSE = encodeSmallInt(0);
   
   int OBJECT_HEADER_SIZE = 2;  // CLASS_DESCRIPTOR + GC_POINTER
 }
