@@ -25,6 +25,7 @@ import static fr.umlv.smalljs.stackinterp.Instructions.RET;
 import static fr.umlv.smalljs.stackinterp.Instructions.STORE;
 import static fr.umlv.smalljs.stackinterp.TagValues.encodeDictObject;
 import static fr.umlv.smalljs.stackinterp.TagValues.encodeSmallInt;
+import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @SuppressWarnings("static-method")
@@ -39,11 +40,11 @@ public class StackInterpreterGCTests {
   }
   private static String execute(Code mainCode, Dictionary dict) {
     var outStream = new ByteArrayOutputStream(8192);
-    var globalEnv = StackInterpreter.createGlobalEnv(new PrintStream(outStream));
+    var globalEnv = StackInterpreter.createGlobalEnv(new PrintStream(outStream, false, UTF_8));
     var mainFun = newMainFunction(mainCode);
     globalEnv.register("main", mainFun);
     StackInterpreter.execute(mainFun, dict, globalEnv);
-    return outStream.toString(StandardCharsets.UTF_8).replace("\r\n", "\n");
+    return outStream.toString(UTF_8).replace("\r\n", "\n");
   }
 
 //  @Tag("Q2") @Test
